requirements.txt
# requirements.txt

# tqdm for progress bars
tqdm

# Pillow for image processing
Pillow